start parser rewrite
change coding styles a bit

parse function now yields an abstract representation of the source code
(except operations, which have to be fixed). this will have to be passed
to the translate function to compile it to bytecode.
when this is done, everything should be back and functioning.

also removed some unnecessary arrays where we could just use pointers.

also also fixed a nasty in find_config